wip: Generate and use targets with usage requirements.
This works with the wip-target-interface branch in my cmake clone
which will soon be under review:

 https://gitorious.org/~steveire/cmake/steveires-cmake

It allows the user to do this:

 find_package(BoostAny)

 add_executable(myexe main.cpp)
 target_link_libraries(myexe boost::any)

and CMake will automatically determine the include directories
needed to use the boost::any library, including any required
dependencies.

This works when using both the build-dir and the install-dir.